[FEAT] UI 구현 및 전반적인 플로우 구현 (#45)
* [FEAT] 향수 상세 페이지 뷰 구현

* [FEAT] 향수 리뷰 UI 구현

* [FEAT] 마이페이지 UI구현 완료

* [FEAT] 홈 로그인 비로그인 로직 구현

* 홈 UI 구현

* [FEAT] 검색 탭 UI 구현

* [FEAT] 비로그인시 향수 좋아요 및 향수 리뷰 제한 구현

* [FEAT] 마이페이지 향수 플로우 구현

* [FEAT] 메뉴탭 닫기 버튼 등 구현

* [FEAT] New label 수정

+++ ScentsNote/Data/Model/Perfume/Response/PerfumeDetailResponseDTO.swift
@@ -15,7 +15,7 @@ struct PerfumeDetailResponseDTO: Decodable, Hashable {
   let abundanceRate: String
   let volumeAndPrice: [String]
   let imageUrls: [String]
-  let score: Float
+  let score: Double
   let seasonal: SeasonalResponseDTO
   let sillage: SillageResponseDTO
   let longevity: LongevityResponseDTO
